Added an option to turn off metadata
I've found that if I do any additional processing to the codelist object (for example, adding a column stating whether the code is active or inactive), the addition of the metadata messes up the export if there are less than 7 codes in the codelist (due to empty rows needing to be created with an unknown number of columns). I thought the easiest and least glitchy way of dealing with this would just be an option to turn metadata off.